requirements.txt
adrf==0.1.9
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
async-property==0.2.2
attrs==25.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
decorator==5.2.1
Django==5.1.6
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-extensions==3.2.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
dotenv==0.9.9
eventlet==0.39.1
executing==2.2.0
flower==2.0.1
greenlet==3.1.1
humanize==4.12.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipython==9.0.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kombu==5.4.2
matplotlib-inline==0.1.7
parso==0.8.4
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
rest-framework-simplejwt==0.0.2
service-identity==24.2.0
setuptools==76.0.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
django-filter==25.1
gunicorn==23.0.0

